Add support for custom highlight groups
This commit allows for themes to add their own highlight groups on top
of the default ones. It does so by iterating over the theme dictionary's
items instead of using g:crystalline_hi_groups. With this commit,
g:crystalline_hi_groups doesn't serve a purpose and can be deleted.

Note that even before this commit, a theme was not required to implement
the default highlight groups anyways: If it didn't have an entry for a
group's name, crystalline#generate_theme() would simply skip generating
a highlight group for this name and continue with the next one.

Also note that, if a user wants to generate separator highlight groups
for the custom groups in a theme, they will need to add mappings to
g:crystalline_supported_sep.