crystalline#left_sep(): Fix argument order
The order for the highlight groups was confusing me a *lot*. In essence,
left_sep() required you to write the group *after* the separator
*first*.

This patch simply changes the order of the arguments to left_sep() and
(hopefully) fixes all uses of left_sep() in the documentation.

The fact that the same highlight group is used for both left_sep() and
right_sep() might trip up some users, so I've also added a note to the
documentation.